Move retraction detail tests to tests_retractions.py.
