+++ api_tests/registrations/views/test_retractions.py
@@ -1,3 +1,4 @@
+from urlparse import urlparse
 from api_tests.nodes.views.test_node_contributors_list import NodeCRUDTestCase
 
 from nose.tools import *  # flake8: noqa
@@ -19,12 +20,15 @@ class TestRetractions(NodeCRUDTestCase):
     def setUp(self):
         super(TestRetractions, self).setUp()
         self.registration = RegistrationFactory(creator=self.user, project=self.public_project)
-        retraction = RetractedRegistrationFactory(registration=self.registration, user=self.registration.creator)
+        self.retraction = RetractedRegistrationFactory(registration=self.registration, user=self.registration.creator)
 
         self.public_pointer_project = ProjectFactory(is_public=True)
         self.public_pointer = self.public_project.add_pointer(self.public_pointer_project,
                                                               auth=Auth(self.user),
                                                               save=True)
+        self.retraction_url = '/{}registrations/{}/'.format(API_BASE, self.registration._id)
+        self.retraction.justification = 'We made a major error.'
+        self.retraction.save()
 
     def test_can_access_retracted_contributors(self):
         url = '/{}registrations/{}/contributors/'.format(API_BASE, self.registration._id)
@@ -106,4 +110,58 @@ class TestRetractions(NodeCRUDTestCase):
         url = '/{}registrations/{}/registrations/'.format(API_BASE, self.registration._id)
         res = self.app.get(url, auth=self.user.auth, expect_errors=True)
         assert_equal(res.status_code, 403)
+
+    def test_retractions_display_limited_fields(self):
+        registration = self.registration
+        res = self.app.get(self.retraction_url, auth=self.user.auth)
+        assert_equal(res.status_code, 200)
+        attributes = res.json['data']['attributes']
+        expected_attributes = {
+            'title': registration.title,
+            'description': registration.description,
+            'date_created': registration.date_created,
+            'date_registered': registration.registered_date,
+            'withdrawal_justification': registration.retraction.justification,
+            'public': None,
+            'category': None,
+            'date_modified': None,
+            'registration': True,
+            'fork': None,
+            'collection': None,
+            'tags': None,
+            'withdrawn': True,
+            'pending_withdrawal': None,
+            'pending_registration_approval': None,
+            'pending_embargo_approval': None,
+            'embargo_end_date': None,
+            'registered_meta': None,
+            'current_user_permissions': None,
+            'registration_supplement': registration.registered_meta.keys()[0]
+        }
+
+        assert_items_equal(attributes, expected_attributes)
+
+        contributors = urlparse(res.json['data']['relationships']['contributors']['links']['related']['href']).path
+        assert_equal(contributors, '/{}registrations/{}/contributors/'.format(API_BASE, registration._id))
+
+        assert_not_in('children', res.json['data']['relationships'])
+        assert_not_in('comments', res.json['data']['relationships'])
+        assert_not_in('node_links', res.json['data']['relationships'])
+        assert_not_in('registrations', res.json['data']['relationships'])
+        assert_not_in('parent', res.json['data']['relationships'])
+        assert_not_in('forked_from', res.json['data']['relationships'])
+        assert_not_in('files', res.json['data']['relationships'])
+        assert_not_in('logs', res.json['data']['relationships'])
+
+    def test_field_specific_related_counts_ignored_if_hidden_field_on_retraction(self):
+        url = '/{}registrations/{}/?related_counts=children'.format(API_BASE, self.registration._id)
+        res = self.app.get(url, auth=self.user.auth)
+        assert_equal(res.status_code, 200)
+        assert_not_in('children', res.json['data']['relationships'])
+
+    def test_field_specific_related_counts_retrieved_if_visible_field_on_retraction(self):
+        url = '/{}registrations/{}/?related_counts=contributors'.format(API_BASE, self.registration._id)
+        res = self.app.get(url, auth=self.user.auth)
+        assert_equal(res.status_code, 200)
+        assert_equal(res.json['data']['relationships']['contributors']['links']['related']['meta']['count'], 1)
